requirements.txt
MouseInfo==0.1.3
mss==9.0.2
numpy==2.1.1
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-Quartz==10.3.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
pytweening==1.2.0
rubicon-objc==0.4.9
